Rename HackTheBox_Academy to HackTheBox_Academy - Network Enumeration with Nmap.md
Module Summary
Nmap is used to identify and scan systems on the network. It is an important part of network diagnostics and evaluation of network-connected systems. In this module, we will learn the basics of this tool and how it can be used efficiently to map out the internal network by identifying live hosts and performing port scanning, service enumeration, and operating system detection.

In this module, we will cover:

An overview of Nmap
Host discovery and port scanning
Saving scan results
Service enumeration
Using the powerful Nmap scripting language
Firewall and IDS/IPS evasion